checkdate and droptables

+++ database/migrations/2017_04_02_140222_eapbs.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use App\Models\Eapb;
 
 class Eapbs extends Migration
 {
@@ -13,17 +14,19 @@ class Eapbs extends Migration
      */
     public function up()
     {
-        Schema::create('eapbs', function(Blueprint $table)
-        {
-            $table->bigInteger('id_entidad', true);
-            $table->string('tipo_entidad', 3)->nullable();
-            $table->string('tipo_identificacion', 3)->nullable();
-            $table->string('num_identificacion', 12)->nullable();
-            $table->string('cod_eapb', 12)->nullable();
-            $table->string('nombre', 300)->nullable();
-            $table->integer('cod_mpio')->nullable();
-            $table->unique(['num_identificacion','cod_eapb'], 'eapbs_uniquekey');
-        });
+        if(!Schema::hasTable('eapbs')){
+            Schema::create('eapbs', function(Blueprint $table)
+            {
+                $table->bigInteger('id_entidad', true);
+                $table->string('tipo_entidad', 3)->nullable();
+                $table->string('tipo_identificacion', 3)->nullable();
+                $table->string('num_identificacion', 12)->nullable();
+                $table->string('cod_eapb', 12)->nullable();
+                $table->string('nombre', 300)->nullable();
+                $table->integer('cod_mpio')->nullable();
+                $table->unique(['num_identificacion','cod_eapb'], 'eapbs_uniquekey');
+            });
+        }
     }
 
     /**
@@ -32,7 +35,7 @@ class Eapbs extends Migration
      * @return void
      */
     public function down()
-    {
-        Schema::drop('eapbs');
+    {   
+        if(Eapb::count() == 0) Schema::drop('eapbs');
     }
 }
